rfactor: renamed game.rs to state.rs
